[None][refactor] generalize MXFP4 checkpoint layout
